services/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.13.3
pydantic-settings==2.5.2
pydantic[email]==2.9.2
passlib[bcrypt]==1.7.4
PyJWT==2.9.0
httpx==0.27.2
docker==7.1.0
celery[redis]==5.4.0
boto3==1.35.42
cryptography==43.0.3
python-multipart==0.0.12
